Tidy up data loader
