+++ README.md
@@ -3,8 +3,8 @@
 Algorithms for matching Mentors to Mentees.
 
 ## TODO
-* [ ] Remove global vars
-* [ ] Add test framework
+* [*] Remove global vars
+* [*] Add test framework
 * [ ] Add tests
 * [ ] add CI
 * [ ] Add license
