+++ src/stable_mentor/__init__.py
@@ -1,9 +1,11 @@
-#    best match for mentors to mentees, mentees win ties
-# Canterbury Tech, Christchurch, NZ
-#  Ian Wells, Calvin Giles
-# takes input from survey monkey to rank mentors and mentees
+"""
+   best match for mentors to mentees, mentees win ties
+Canterbury Tech, Christchurch, NZ
+ Ian Wells, Calvin Giles
+takes input from survey monkey to rank mentors and mentees
 
-'''
+
+Based on:
 @Author: Ryan Schachte
 @Publication-Date: 1/12/17 5:23 PM
 @Description:
@@ -16,39 +18,8 @@ We can define "matching" and "stable" by the following definitions.
 Matching: Mapping from the elements of one set to the elements of another set
 Stable: No element A of the first set that prefers an element B of the second set over its current partner
         such that element B prefers element A over its current partner.
-'''
-
-import pandas
-import collections
-
-
-
-def get_sorted_rankings(rankings):
-    my_sorted_rankings = rankings.sort_values(ascending=True)
-    return [
-        person_name
-        for person_name, rank
-        in my_sorted_rankings.iteritems()
-    ]
-
-
-#The mentors that the mentee prefers
-#preferred_rankings_mentor = {
-
-#	'lizzy': 	['ryan', 'blake', 'josh', 'connor'],
-#	'sarah': 	['ryan', 'blake', 'connor', 'josh'],
-#	'zoey':  	['connor', 'josh', 'ryan', 'blake'],
-#	'daniella':	['ryan', 'josh', 'connor', 'blake']
-#}
-
-
-#The mentees that the mentor prefers
-#preferred_rankings_mentee = {
-#	'ryan': 	['lizzy', 'sarah', 'zoey', 'daniella'],
-#	'josh': 	['sarah', 'lizzy', 'daniella', 'zoey'],
-#	'blake': 	['sarah', 'daniella', 'zoey', 'lizzy'],
-#	'connor': 	['lizzy', 'sarah', 'zoey', 'daniella']
-#}
+"""
+from stable_mentor.data_loader import load_data
 
 
 def init_free_mentor(preferred_rankings_mentor):
@@ -61,6 +32,7 @@ def init_free_mentor(preferred_rankings_mentor):
         free_mentor.append(mentor)
 
     return free_mentor
+
 
 def begin_matching(preferred_rankings_mentor, preferred_rankings_mentee, free_mentor, tentative_engagements, mentor):
     '''Find the first free mentee available to a mentor at
@@ -114,48 +86,3 @@ def stable_matching(preferred_rankings_mentor, preferred_rankings_mentee):
             tentative_engagements = begin_matching(preferred_rankings_mentor, preferred_rankings_mentee, free_mentor, tentative_engagements, mentor)
     print(free_mentor)
     return tentative_engagements
-
-
-def load_data(data_path):
-    return pandas.read_excel(data_path, header=[0, 1])
-
-
-def parse_data(df):
-    preferred_rankings_mentee = {}
-    preferred_rankings_mentor = {}
-
-    for i, row in df.iterrows():
-        name = row["Select your name"].iloc[0]  # change this for production
-        yes_mentee = row["Are you a mentor or mentee?"].iloc[1]
-        yes_mentor = row["Are you a mentor or mentee?"].iloc[0]
-
-        if (yes_mentor == "Mentor"):
-            print("I am a mentor", name, "\n")
-            my_mentees = row["Rank your choices of mentee from 1 ( most preferable) to least"]
-            preferred_rankings_mentee[name] = get_sorted_rankings(my_mentees)
-            print("my mentees:", i, name, "\n", my_mentees, "\n", preferred_rankings_mentee)
-        elif (yes_mentee == "Mentee"):
-            print("I am a mentee", name)
-            my_mentors = row["Rank your choices of mentor from 1 ( most preferable) to least"]
-            my_sorted_mentors = my_mentors.sort_values(ascending=True)
-            preferred_rankings_mentor[name] = get_sorted_rankings(my_mentors)
-
-            print("my mentors:", i, name, "\n", my_mentors, "\n", preferred_rankings_mentor)
-        else:
-            print("error in spreadsheet")
-
-    return preferred_rankings_mentor, preferred_rankings_mentee
-
-
-def main():
-    df = load_data("../data/testdata9.xlsx")
-    preferred_rankings_mentor, preferred_rankings_mentee = parse_data(df)
-    print(f"PREFERRED_RANKINGS_MENTOR: {preferred_rankings_mentor}")
-    print(f"PREFERRED_RANKINGS_MENTEE: {preferred_rankings_mentee}")
-
-    tentative_engagements = stable_matching(preferred_rankings_mentor, preferred_rankings_mentee)
-    print("BEST MATCHES" , tentative_engagements)
-
-
-if __name__ == "__main__":
-    main()
